requirements.txt
aiofiles==24.1.0
aiogoogle==5.13.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
frozenlist==1.5.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
idna==3.10
multidict==6.1.0
numpy==2.1.3
oauthlib==3.2.2
pandas==2.2.3
propcache==0.2.1
pyasn1==0.6.1
pyasn1-modules==0.4.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tonyg-rfc3339==0.1
tzdata==2024.2
urllib3==2.2.3
yarl==1.18.3
